fix: remove svg rendering bug
This fix adjusts the introduction illustration (with the peacock, duck, and antelope) so that it no longer displays incorrectly in Safari or Edge. Previously, the introduction partial loaded an svg as an image — and the svg's viewbox included subpixels. This fix adds a Handlebars partial that includes the svg ccode, then adds the partial to the introduction partial, so that the image is loaded as an inline svg but no-one has to scan through the entire SVG declaration to make an adjustment to the introduction component code.