refactor: fluid hero component
This commit refactors the hero component on the AME website to better utilize fluid units (vw and vh) to reposition the Zebra in a way that helps visitors see the zebra and the text together. Previously,
the Zebra would bogart the focus of the container when it took up inline space in the component at mobile sizes.