refactor: optimize illustration svg files
This commit cleans up illustrations included in the A Modern Eden Project. It fixes the bug associated with non-perfectly tiled backgrounds. It also re-optimizes (by running through SVGOMG) many of the individual objects included in the illustrated scenes, excluding the fish – which will be included in a separate animation commit.