fix: remove persistent cracks in svg backgrounds
This commit fixes the persistent cracks in the tiled svg backgrounds present on the page. Previously, these cracks were attempted to be fixed by setting an explicit height and width on each respective svg background. As it turns out, this was only part of the problem. According to the MDN spec, svg background images are sized using an explicit or an implicit size — and that one wins out. So as it turns out, I was attempting to set an explicit size on the svgs but then setting background-sizing properties that overrule that size. This has been fixed by removiong the background-size css properties, adjusting the explicit size of the svg backgrounds as needed, and also including an additional small-screen only svg background as needed.